(failing) get compiler errors to where it's trying to get line numbers
