+++ test/integration/assets/compile_error/undeclared_identifier.exs
@@ -1,0 +1,10 @@
+defmodule ZiglerTest.CompileError.InvalidIdentifier do
+  use Zigler
+
+  ~Z"""
+  /// nif: foo/1
+  fn foo(v: i64) i64 {
+    return not_v;
+  }
+  """
+end
