--- test/integration_legacy/test_helper.exs
@@ -1,5 +1,0 @@
-Logger.configure(level: :warn)
-
-File.rm_rf!("priv/nifs")
-
-ExUnit.start()
